+++ Apps/Monolith/Modules/Products/Domain/CreateOne/Dtos/ICreateOneProductDto.cs
@@ -15,8 +15,8 @@ public interface ICreateOneVariantDto
 {
     public Guid Id { get; init; }
     public string Name { get; init; }
-    public float Price { get; init; }
-    public float Cost { get; init; }
+    public decimal Price { get; init; }
+    public decimal Cost { get; init; }
     public string CreatedAtTimezone { get; init; }
     public Guid CreatedBy { get; init; }
     public Guid ProductId { get; init; }
