+++ Apps/Monolith/Modules/Variants/Domain/CreateOne/Inputs/ICreateOneVariantInput.cs
@@ -6,8 +6,8 @@ public interface ICreateOneVariantInput<T>
 {
     public Optional<Guid> Id { get; set; }
     public string Name { get; set; }
-    public float Price { get; set; }
-    public float Cost { get; set; }
+    public decimal Price { get; set; }
+    public decimal Cost { get; set; }
     public DateTimeZone CreatedAtTimezone { get; set; }
     public Guid CreatedBy { get; set; }
     public Guid ProductId { get; set; }
