+++ Apps/Monolith/Modules/Variants/Domain/UpdateOne/Dtos/IUpdateOneVariantDto.cs
@@ -4,8 +4,8 @@ public interface IUpdateOneVariantDto
 {
     public Guid Id { get; init; }
     public string? Name { get; init; }
-    public float? Price { get; init; }
-    public float? Cost { get; init; }
+    public decimal? Price { get; init; }
+    public decimal? Cost { get; init; }
     public string UpdatedAtTimezone { get; init; }
     public Guid UpdatedBy { get; init; }
 }
