+++ Apps/Monolith/Modules/Variants/Domain/UpdateOne/Inputs/IUpdateOneVariantInput.cs
@@ -6,8 +6,8 @@ public interface IUpdateOneVariantInput
 {
     public Guid Id { get; set; }
     public string? Name { get; set; }
-    public float? Price { get; set; }
-    public float? Cost { get; set; }
+    public decimal? Price { get; set; }
+    public decimal? Cost { get; set; }
     public DateTimeZone UpdatedAtTimezone { get; set; }
     public Guid UpdatedBy { get; set; }
 }
